Automatically start and stop embedded postgres.
RMB automatically starts embedded postgres if needed.
It is not needed if DB_ environment variables point to
a running postgres. A developer can save much time by
providing a running database and can inspect the database
afterwards.
Stop embedded postgres after all IT tests have run (this does
not stop a database the DB_ variables point to).
Delete the test tenant in the database at the start of each IT
test if needed.
This also works if the developers runs a single IT test class only.